backend/user-suppliers/requirements.txt
psycopg2-binary==2.9.9
openai==1.12.0
requests==2.31.0
beautifulsoup4==4.12.3